taskmanager/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
python-dateutil==2.8.2
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9